tests/html-dom.rs: explain `doc_render_bug`
