+++ src/dom_tree.rs
@@ -2,6 +2,7 @@ mod helpers;
 mod ops;
 mod traversal;
 mod tree;
+mod validation;
 
 pub use ops::TreeNodeOps;
 pub use traversal::Traversal;
