+++ tests/html-dom.rs
@@ -175,7 +175,8 @@ fn doctype() {
 
 #[cfg_attr(not(target_arch = "wasm32"), test)]
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
-fn doc_render_bug() {
+#[should_panic]
+fn test_issue_node_append_the_next_sibling() {
     const HTML: &str = r#"<!DOCTYPE html><html><head></head><body>
            <div id="parent">
                <div id="child" class="child">Child</div>
@@ -187,17 +188,42 @@ fn doc_render_bug() {
     let sel = doc.select("#child");
     let child = sel.nodes().first().unwrap();
 
-    // Create a wrapper directly in the same tree
+    // - Create a wrapper directly in the same tree area.
+    // This element is detached from the DOM tree -- it has no parent.
     let wrapper = doc.tree.new_element("div");
     wrapper.set_attr("id", "wrapper");
 
-    // Insert wrapper before child in the parent
+    // - Insert wrapper before existing child in the parent. 
+    // Under the hood `insert_before_of` will call `remove_from_parent` for the new children,
+    // because we don`t know if it is a new node without bound to the tree or not.
+    // At this point `child` becomes a next sibling of the `wrapper`.
     doc.tree.insert_before_of(&child.id, &wrapper.id);
 
-    // Move child into wrapper as the only child
+    // If we don't call `remove_from_parent` for the `child` node,
+    // it will still be the next sibling of the `wrapper` node
+    // and at the same time also a child of the `wrapper` node.
+    // In this case, it will be both the first and the last child of the `wrapper`.
+    // - Move child into wrapper as the only child
+    //doc.tree.remove_from_parent(&child.id);
     doc.tree.append_child_of(&wrapper.id, &child.id);
+    // This may seem inconvenient, but the `Tree` methods are intended more for internal use,
+    // while `NodeRef` is designed for external usage.
 
-    // Normalize the document for printing
+    // So yes, without "detaching" the child node from the tree, the `Tree` instance becomes invalid.
+
+    // **But I cannot consider this a bug**.
+
+    // Because there are ways to avoid this situation:
+    // 1. Using `Node::insert_before` and `Node::append_child` methods, which will do the dirty work for you:
+    // child.insert_before(&wrapper);
+    // wrapper.append_child(child);
+
+    // 2. Using `Node::replace_with` and `Node::append_child` methods, which also handle cleanup internally.
+    // But note that the child node will be removed from the tree twice in this case.
+    // child.replace_with(&wrapper);
+    // wrapper.append_child(child);
+
+    // - Normalize the document for printing
     doc.normalize();
 
     // Check to see if the structure is valid
